Touchscree/ekft3k: Auto retry 5 times on i2c reads
The i2c devices we are connecting to may not be ready yet, or just
busy. Retry with a 5ms pause in order to still obtain i2c data.
This improves overall i2c data robustness

Based on commit of  Dennis Rassmann <showp1984@gmail.com>
Modified to 5 to more faster.